add main.js, format files
